Manager Refactor
