--- src/main/java/io/github/sefiraat/crystamaehistoria/listeners/ListenerManager.java
@@ -1,31 +1,0 @@
-package io.github.sefiraat.crystamaehistoria.listeners;
-
-import io.github.sefiraat.crystamaehistoria.CrystamaeHistoria;
-import org.bukkit.event.Listener;
-
-public class ListenerManager {
-
-    public ListenerManager() {
-        addListener(new ArmorStandInteract());
-        addListener(new SpellCastListener());
-        addListener(new SpellEffectListener());
-        addListener(new CrystalBreakListener());
-        addListener(new BlockRemovalListener());
-        addListener(new MaintenanceListener());
-        addListener(new RefractingLensListener());
-        addListener(new ThaumaturgicSaltsListener());
-        addListener(new CrystaDowngradeListener());
-        addListener(new SatchelListener());
-        addListener(new EndermanInhibitorListener());
-        addListener(new MobCandleListener());
-        addListener(new DisplayItemListener());
-        addListener(new PoseChangerListener());
-        addListener(new PhilosophersSprayListener());
-        addListener(new MiscListener());
-    }
-
-    private void addListener(Listener listener) {
-        CrystamaeHistoria.getPluginManager().registerEvents(listener, CrystamaeHistoria.getInstance());
-    }
-
-}
